+++ src/store/store.ts
@@ -1,7 +1,8 @@
 import { configureStore } from "@reduxjs/toolkit";
+import userReducer from "@/data/userdataslice";
 
 export const store = configureStore({
-  reducer: {},
+  reducer: { userReducer },
 });
 
 // Infer the `RootState` and `AppDispatch` types from the store itself
